requirements.txt
Django~=4.1.5
djangorestframework~=3.14.0
djoser~=2.1.0
xhtml2pdf~=0.2.9
django-filter~=22.1
social-auth-core>=4.3,<5.0
social-auth-app-django>=5.0,<6.0